# api_documentation.yaml
openapi: 3.0.3
info:
  title: SmartCampus API
  version: 1.0.0
  description: API pour la gestion des utilisateurs, rôles, programmes et modules.
paths:
...
  /api/etudiants/:
    get:
      operationId: etudiants_list
      tags:
      - etudiants
      security:
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Etudiant'
          description: ''
    post:
      operationId: etudiants_create
      tags:
      - etudiants
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EtudiantRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/EtudiantRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/EtudiantRequest'
        required: true
      security:
      - {}
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Etudiant'
          description: ''
  /api/etudiants/{id}/:
    get:
      operationId: etudiants_retrieve
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      tags:
      - etudiants
      security:
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Etudiant'
          description: ''
    put:
      operationId: etudiants_update
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      tags:
      - etudiants
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EtudiantRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/EtudiantRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/EtudiantRequest'
        required: true
      security:
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Etudiant'
          description: ''
    patch:
      operationId: etudiants_partial_update
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      tags:
      - etudiants
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedEtudiantRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedEtudiantRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedEtudiantRequest'
      security:
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Etudiant'
          description: ''
    delete:
      operationId: etudiants_destroy
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      tags:
      - etudiants
      security:
      - {}
      responses:
        '204':
          description: No response body
...
  /api/notes/{id}/details/:
    get:
      operationId: notes_details_retrieve
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this note.
        required: true
      tags:
      - notes
      security:
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Note'
          description: ''
  /api/notes/par-filiere/:
    get:
      operationId: notes_par_filiere_retrieve
      tags:
      - notes
      security:
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Note'
          description: ''
  /api/notes/par-filiere-niveau/:
    get:
      operationId: notes_par_filiere_niveau_retrieve
      tags:
      - notes
      security:
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Note'
          description: ''
  /api/notes/par-module/:
    get:
      operationId: notes_par_module_retrieve
      tags:
      - notes
      security:
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Note'
          description: ''
...
  /api/v2/notes/{id}/details/:
    get:
      operationId: v2_notes_details_retrieve
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this note.
        required: true
      tags:
      - v2
      security:
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcademicNote'
          description: ''
...
  /api/v2/notes/par-filiere-niveau/:
    get:
      operationId: v2_notes_par_filiere_niveau_retrieve
      tags:
      - v2
      security:
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcademicNote'
          description: ''
  /api/v2/notes/par-module/:
    get:
      operationId: v2_notes_par_module_retrieve
      tags:
      - v2
      security:
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcademicNote'
          description: ''
...
components:
  schemas:
...
    AcademicNote:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        etudiant_nom:
          type: string
          readOnly: true
        evaluation_nom:
          type: string
          readOnly: true
        classe_nom:
          type: string
          readOnly: true
        module_nom:
          type: string
          readOnly: true
        session:
          $ref: '#/components/schemas/SessionEnum'
        annee_academique:
          type: string
          maxLength: 9
        note_cc:
          type: string
          format: decimal
          pattern: ^-?\d{0,3}(?:\.\d{0,2})?$
          nullable: true
          description: Note CC sur 20
        note_sn:
          type: string
          format: decimal
          pattern: ^-?\d{0,3}(?:\.\d{0,2})?$
          nullable: true
          description: Note SN sur 20
        note_tp:
          type: string
          format: decimal
          pattern: ^-?\d{0,3}(?:\.\d{0,2})?$
          nullable: true
          description: Note TP sur 20 (si applicable)
        note_rattrapage:
          type: string
          format: decimal
          pattern: ^-?\d{0,3}(?:\.\d{0,2})?$
          nullable: true
          description: Note Rattrapage sur 20 (facultatif)
        note_finale:
          type: string
          format: decimal
          pattern: ^-?\d{0,3}(?:\.\d{0,2})?$
          nullable: true
          description: Note finale sur 20
        etudiant:
          type: integer
        module:
          type: integer
        classe:
          type: integer
          nullable: true
        evaluation:
          type: integer
          nullable: true
      required:
      - classe_nom
      - etudiant
      - etudiant_nom
      - evaluation_nom
      - id
      - module
      - module_nom
      - session
...
    Etudiant:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        matricule:
          type: string
          maxLength: 30
        nom:
          type: string
          maxLength: 100
        date_naissance:
          type: string
          format: date
          nullable: true
        contact:
          type: string
          maxLength: 50
        email:
          type: string
          format: email
          maxLength: 254
        filiere:
          allOf:
          - $ref: '#/components/schemas/Formation'
          readOnly: true
        statut:
          $ref: '#/components/schemas/EtudiantStatutEnum'
        documents:
          type: array
          items:
            $ref: '#/components/schemas/EtudiantDocument'
          readOnly: true
        inscriptions:
          type: array
          items:
            $ref: '#/components/schemas/Inscription'
          readOnly: true
      required:
      - contact
      - documents
      - email
      - filiere
      - id
      - inscriptions
      - nom
    EtudiantDocument:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        type_document:
          type: string
          nullable: true
          maxLength: 50
        fichier:
          type: string
          format: uri
        date_upload:
          type: string
          format: date-time
          readOnly: true
      required:
      - date_upload
      - fichier
      - id
...
    EtudiantRequest:
      type: object
      properties:
        matricule:
          type: string
          maxLength: 30
        nom:
          type: string
          minLength: 1
          maxLength: 100
        date_naissance:
          type: string
          format: date
          nullable: true
        contact:
          type: string
          minLength: 1
          maxLength: 50
        email:
          type: string
          format: email
          minLength: 1
          maxLength: 254
        filiere_id:
          type: integer
          writeOnly: true
        niveau_id:
          type: integer
          writeOnly: true
        classe_id:
          type: integer
          writeOnly: true
        annee_academique_ref_id:
          type: integer
          writeOnly: true
        statut:
          $ref: '#/components/schemas/EtudiantStatutEnum'
      required:
      - contact
      - email
      - filiere_id
      - nom
    EtudiantStatutEnum:
      enum:
      - Pré-inscrit
      - Inscrit
      type: string
      description: |-
        * `Pré-inscrit` - Pré-inscrit
        * `Inscrit` - Inscrit
...
    Formation:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        intitule:
          type: string
          maxLength: 200
      required:
      - id
      - intitule
...
    Inscription:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        etudiant:
          type: integer
        classe:
          type: integer
          nullable: true
        classe_nom:
          type: string
          readOnly: true
        niveau:
          type: integer
        niveau_nom:
          type: string
          readOnly: true
        formation_nom:
          type: string
          readOnly: true
        annee_academique:
          type: string
          description: 'Ex: 2024-2025'
          maxLength: 9
        annee_academique_ref:
          type: integer
          nullable: true
        annee_academique_ref_libelle:
          type: string
          readOnly: true
        date_inscription:
          type: string
          format: date
          readOnly: true
      required:
      - annee_academique
      - annee_academique_ref_libelle
      - classe_nom
      - date_inscription
      - etudiant
      - formation_nom
      - id
      - niveau
      - niveau_nom
...
    Note:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        etudiant:
          type: integer
        etudiant_nom:
          type: string
          readOnly: true
        etudiant_matricule:
          type: string
          readOnly: true
        formation:
          type: string
          readOnly: true
        classe:
          type: integer
          nullable: true
        classe_nom:
          type: string
          readOnly: true
        module:
          type: integer
        module_nom:
          type: string
          readOnly: true
        evaluation:
          type: integer
          nullable: true
        evaluation_nom:
          type: string
          readOnly: true
        session:
          $ref: '#/components/schemas/SessionEnum'
        note_cc:
          type: string
          format: decimal
          pattern: ^-?\d{0,3}(?:\.\d{0,2})?$
          nullable: true
          description: Note CC sur 20
        note_sn:
          type: string
          format: decimal
          pattern: ^-?\d{0,3}(?:\.\d{0,2})?$
          nullable: true
          description: Note SN sur 20
        note_tp:
          type: string
          format: decimal
          pattern: ^-?\d{0,3}(?:\.\d{0,2})?$
          nullable: true
          description: Note TP sur 20 (si applicable)
        note_rattrapage:
          type: string
          format: decimal
          pattern: ^-?\d{0,3}(?:\.\d{0,2})?$
          nullable: true
          description: Note Rattrapage sur 20 (facultatif)
        note_finale:
          type: string
          format: decimal
          pattern: ^-?\d{0,3}(?:\.\d{0,2})?$
          nullable: true
          description: Note finale sur 20
        note_sur_20:
          type: string
          readOnly: true
        besoin_rattrapage:
          type: string
          readOnly: true
      required:
      - besoin_rattrapage
      - classe_nom
      - etudiant
      - etudiant_matricule
      - etudiant_nom
      - evaluation_nom
      - formation
      - id
      - module
      - module_nom
      - note_sur_20
      - session
...
    PatchedEtudiantRequest:
      type: object
      properties:
        matricule:
          type: string
          maxLength: 30
        nom:
          type: string
          minLength: 1
          maxLength: 100
        date_naissance:
          type: string
          format: date
          nullable: true
        contact:
          type: string
          minLength: 1
          maxLength: 50
        email:
          type: string
          format: email
          minLength: 1
          maxLength: 254
        filiere_id:
          type: integer
          writeOnly: true
        niveau_id:
          type: integer
          writeOnly: true
        classe_id:
          type: integer
          writeOnly: true
        annee_academique_ref_id:
          type: integer
          writeOnly: true
        statut:
          $ref: '#/components/schemas/EtudiantStatutEnum'
...
    SessionEnum:
      enum:
      - Semestre 1
      - Semestre 2
      type: string
      description: |-
        * `Semestre 1` - Semestre 1
        * `Semestre 2` - Semestre 2
...